% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean downloaded data frame}
\usage{
eq_clean_data(filename)
}
\arguments{
\item{filename}{A character vector of the name of the tab-delimited file to be read.}
}
\value{
This function returns a clean data frame.
}
\description{
This function reads an Earthquake dataset in a tab-delimited file to clean some of the variables.
The function first checks whether the file to be read actually exists.If it does not exist, then it downloads from the NOOA website.
Then, the function adds a date column created by uniting the year, month, day of each earthquake and converting it to the Date class.
Finally, the function converts the LATITUDE and LONGITUDE columns to numeric class.
}
\examples{
\dontrun{
eq_clean_data("signif.txt")
}

}
