% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean up LOCATION_NAME column from downloaded data frame}
\usage{
eq_location_clean(EarthquakeData)
}
\arguments{
\item{EarthquakeData}{A dataframe.}
}
\value{
This function returns a data frame with the column LOCATION_NAME cleaned.
}
\description{
This function reads an Earthquake dataset in a dataframe format to clean the LOCATION_NAME column by stripping out the country name (including the colon) and convert names to title case (as opposed to all caps).
}
\examples{
\dontrun{
eq_location_clean(EarthquakeDataFrame)
eq_clean_data("signif.txt") \%>\% eq_location_clean()
}

}
