% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapping_functions.R
\name{eq_map}
\alias{eq_map}
\title{Function to visualize the epicenters of earthquakes with annotation in a leaflet interactive map}
\usage{
eq_map(dataframe, annot_col)
}
\arguments{
\item{dataframe}{A filtered dataframe with the earthquakes to visualize.}

\item{annot_col}{A character string indicating the column used for the annotation in the pop-up window.}
}
\value{
A leaflet fully interactive map plotting the epicenters of selected earthquakes.
}
\description{
This function reads a filtered data frame with earthquakes to visualize. The function maps the epicenters (LATITUDE/LONGITUDE) and annotates each point in a pop-up window containing annotation data stored in a column of the data frame. The user is able to choose which column is used for the annotation in the pop-up with the function argument named annot_col.
 Each earthquake is shown with a circle, and the radius of the circle is proportional to the earthquake's magnitude.
}
\examples{
\dontrun{
eq_clean_data("signif.txt") \%>\% dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATETIME) >= 2000) \%>\% eq_map(annot_col = "DATETIME")
}

}
