% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geom_functions.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Function to display the earthquake timelines}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
The function displays timelines of earthquakes based on the \code{GeomTimeline}.
}
