% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geom_functions.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Function to display the annotated labels for the earthquake timelines}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, fontsize = 10, angle = 45, ...)
}
\description{
The function displays the annotated labels for the timelines of earthquakes based on the \code{GeomTimelineLabel}.
}
