% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geom_functions.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{ggplot geom to plot a time line of earthquakes}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\arguments{
\item{x}{A character vector of the dates when the earthquakes took place.}

\item{y}{Optional aesthetic which contains a character vector of the country names, wherein the earthquakes took place.}

\item{fill}{Optional aesthetic which contains a numeric vector of the number of deaths in each earthquake, which will define the colour of the points to be plotted.}

\item{size}{Optional aesthetic which contains a numeric vector of the magnitude (in the Richter scale) of each earthquake, which will define the size of the points to be plotted.}

\item{alpha}{Optional aesthetic that represents a number between 0 and 1, which indicates the level of transparency of represented points.}
}
\value{
An earthquakes' timeline geom which can be added to a ggplot.
}
\description{
This function reads an Earthquake dataset in a dataframe format to plot a time line of earthquakes ranging from the minimum to the maximum dates with a point for each earthquake.
The x aesthetic is a date and an optional y aesthetic is a factor that stratifies the data by country, in which case multiple time lines will be plotted, one for each country.
In addition to showing the dates on which the earthquakes occur, it is also possible to show the magnitudes (i.e. Richter scale value) and the number of deaths associated with each earthquake.
}
\examples{
\dontrun{
quakedata<-eq_clean_data("signif.txt")
ggplot() + geom_timeline(data = quakedata,aes(x = DATETIME, y = COUNTRY, size = EQ_PRIMARY, fill = TOTAL_DEATHS)) + scale_color_gradient()
}

}
\keyword{datasets}
